move biomeutil stuff into common
