+++ platforms/forge/src/main/java/com/dfsek/terra/forge/mixin/lifecycle/DataPackContentsMixin.java
@@ -2,6 +2,8 @@ package com.dfsek.terra.forge.mixin.lifecycle;
 
 import com.dfsek.terra.forge.util.BiomeUtil;
 import com.dfsek.terra.forge.util.TagUtil;
+import com.dfsek.terra.mod.util.MinecraftUtil;
+
 import net.minecraft.server.DataPackContents;
 import net.minecraft.util.registry.DynamicRegistryManager;
 import net.minecraft.util.registry.Registry;
@@ -23,6 +25,6 @@ public class DataPackContentsMixin {
         
         Registry<Biome> biomeRegistry = dynamicRegistryManager.get(Registry.BIOME_KEY);
         TagUtil.registerBiomeTags(biomeRegistry);
-        BiomeUtil.registerFlora(biomeRegistry);
+        MinecraftUtil.registerFlora(biomeRegistry);
     }
 }
